+++ code/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [#493](https://github.com/InditexTech/weavejs/issues/493) No automatic exit when the user switches to another tool when in crop mode
+- [#500](https://github.com/InditexTech/weavejs/issues/500) Textarea not removed in microfrontend; document.getElementById fails
 
 ## [0.38.0] - 2025-07-14